+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Braintree iOS SDK Release Notes
 
+## 3.2.0-rc3 (2014-08-14)
+
+* Add support for app switch configuration via client token.
+
 ## 3.2.0-rc2 (2014-08-13)
 
 * Fixed the `payPalButtonDidCancel:` delegate method to take button as parameter (breaking change)
@@ -10,6 +14,17 @@
 * Add support and docs for PayPal Touch app switch
 * Known Issues:
   * Drop In view controller is not yet compatible with PayPal Touch (app switch auth)
+
+## 3.1.1 (2014-08-17)
+
+* Enhancements
+  * Accept four digit years in expiry field
+  * Internationalize
+  * Support iOS 8 SDK
+* Integration changes
+  * Merge `api` and `API` directory content
+  * Deprecate `savePaypalPaymentMethodWithAuthCode:correlationId:success:failure` in favor of
+    `savePaypalPaymentMethodWithAuthCode:applicationCorrelationID:success:failure`
 
 ## 3.1.0 (2014-07-22)